64bit now working

+++ include/Pyros3D/Ext/StringIDs/StringID.hpp
@@ -12,7 +12,7 @@
 
 namespace p3d
 {
-	typedef unsigned long StringID;
+	typedef uint32 StringID;
 
 	StringID MakeStringID(const std::string &Name);
 	StringID MakeStringIDFromChar(const uchar* data, const uint32 &length);
